🔨 refactor: instance, interceptors 분리
