+++ src/api/axiosInstance/index.ts
@@ -1,4 +1,6 @@
+import axios from 'axios';
 import { ERROR_MESSAGE, HTTP_STATUS_CODE } from '@/constants/api';
+import { responseError, setToken } from './interceptors';
 
 export type Message =
   | (typeof ERROR_MESSAGE)[keyof typeof ERROR_MESSAGE]
@@ -11,3 +13,12 @@ export interface ErrorResponseData {
   statusCode?: HttpStatusCode;
   message?: Message;
 }
+
+export const axiosInstance = axios.create({
+  baseURL: import.meta.env.VITE_BASE_URL,
+  authorization: true
+});
+
+axiosInstance.interceptors.request.use(setToken);
+
+axiosInstance.interceptors.response.use((response) => response, responseError);
